Move widgetGeneratorAction to JobController for coherence
