+++ app/Resources/views/admin/index.html.twig
@@ -93,9 +93,6 @@
                 <a href="{{ path("process_update_list") }}" class="waves-effect waves-light btn lime darken-1">
                     <i class="material-icons left">assignment</i>Gérer les nouveautés
                 </a>
-                <a href="{{ path("shift_widget_generator") }}" class="waves-effect waves-light btn teal">
-                    <i class="material-icons left">tune</i>Générer un widget (créneau)
-                </a>
                 <a href="{{ path("dynamic_content_list") }}" class="waves-effect waves-light btn teal">
                     <i class="material-icons left">content_paste</i>Contenus dynamiques
                 </a>
