+++ app/Resources/views/admin/job/list.html.twig
@@ -69,5 +69,10 @@
     </table>
 
     <br />
-    <a href="{{ path('job_new') }}" class="btn"><i class="material-icons left">add</i>Ajouter un poste</a>
+    <a href="{{ path('job_new') }}" class="btn">
+        <i class="material-icons left">add</i>Ajouter un poste
+    </a>
+    <a href="{{ path("job_widget_generator") }}" class="btn">
+        <i class="material-icons left">tune</i>Générer un widget (créneaux d'un poste)
+    </a>
 {% endblock %}
